KiteUtils

Documentation for KiteUtils.

KiteUtils.SysState

KiteUtils.azimuth_east

KiteUtils.calc_elevation

KiteUtils.demo_state

KiteUtils.ground_dist

KiteUtils.rot

KiteUtils.rot3d

KiteUtils.SysState — Type

SysState{P}

Basic system state. One of these is saved per time step. P is the number of tether particles.

time::Float64

time since start of simulation in seconds

orient::StaticArrays.MVector{4, Float32}

orientation of the kite (quaternion)

elevation::Float32

elevation angle in radians

azimuth::Float32

azimuth angle in radians

l_tether::Float32

tether length [m]

v_reelout::Float32

reel out velocity [m/s]

force::Float32

tether force [N]

depower::Float32

depower settings

v_app::Float32

norm of apparent wind speed [m/s]

X::StaticArrays.MVector{P, Float32} where P

vector of particle positions in x

Y::StaticArrays.MVector{P, Float32} where P

vector of particle positions in y

Z::StaticArrays.MVector{P, Float32} where P

vector of particle positions in z

KiteUtils.azimuth_east — Method

function azimuth_east(vec)

Calculate the azimuth angle in radian from the kite position in ENU reference frame. Zero east. Positive direction clockwise seen from above. Valid range: -π .. π.

KiteUtils.calc_elevation — Method

 function calc_elevation(vec)

Calculate the elevation angle in radian from the kite position.

KiteUtils.demo_state — Function

function demo_state(P, height=6.0, time=0.0)

Create a demo state with a given height and time. P is the number of tether particles.

Returns a SysState instance.

KiteUtils.ground_dist — Method

function ground_dist(vec)

Calculate the ground distance of the kite from the groundstation based on the kite position (x,y,z, z up).

KiteUtils.rot — Method

rot(pos_kite, pos_before, v_app)

Calculate the rotation matrix of the kite based on the position of the last two tether particles and the apparent wind speed vector.

KiteUtils.rot3d — Method

rot3d(ax, ay, az, bx, by, bz)

Calculate the rotation of reference frame (ax, ay, az) so that it matches the reference frame (bx, by, bz). All parameters must be 3-element vectors. Both refrence frames must be orthogonal, all vectors must already be normalized.

Source: TRIAD_Algorithm