Exported Functions

Reading config files

KiteUtils.set_data_path — Method

set_data_path(data_path)

Set the directory for log and config files.

KiteUtils.load_settings — Function

load_settings(project="")

Load the project with the given file name. The default project is determined by the content of the file system.yaml .

The project must include the path and the suffix .yaml .

KiteUtils.copy_settings — Method

copy_settings()

Copy the default settings.yaml and system.yaml files to the folder DATAPATH (it will be created if it doesn't exist).

KiteUtils.se — Method

se()

Getter function for the Settings struct.

The default project is determined by the content of the file system.yaml .

Also look at the default example: settings.yaml .

Creating test data

KiteUtils.demo_state — Function

function demo_state(P, height=6.0, time=0.0)

Create a demo state with a given height and time. P is the number of tether particles.

Returns a SysState instance.

KiteUtils.demo_syslog — Function

function demo_syslog(P, name="Test flight"; duration=10)

Create a demo flight log with given name [String] and duration [s] as StructArray. P is the number of tether particles.

KiteUtils.demo_log — Function

function demo_log(P, name="Test_flight"; duration=10)

Create an artifical SysLog struct for demonstration purposes. P is the number of tether particles.

Loading, saving and converting log files

KiteUtils.load_log — Method

function load_log(P, filename::String)

Read a log file that was saved as .arrow file. P is the number of tether particles.

KiteUtils.save_log — Method

function save_log(flight_log)

Save a fligh log of type SysLog as .arrow file. P is the number of tether particles.

The function set_data_path(data_path) can be used to set the directory for the log files.

Rotation matrices

KiteUtils.rot3d — Method

rot3d(ax, ay, az, bx, by, bz)

Calculate the rotation matrix that needs to be applied on the reference frame (ax, ay, az) to match the reference frame (bx, by, bz). All parameters must be 3-element vectors. Both refrence frames must be orthogonal, all vectors must already be normalized.

Source: TRIAD_Algorithm

KiteUtils.rot — Method

rot(pos_kite, pos_before, v_app)

Calculate the rotation matrix of the kite based on the position of the last two tether particles and the apparent wind speed vector.

Geometric calculations

Calculate the elevation angle, the azimuth angle and the ground distance based on the kite position.

KiteUtils.calc_elevation — Method

 function calc_elevation(vec)

Calculate the elevation angle in radian from the kite position.

KiteUtils.azimuth_east — Method

function azimuth_east(vec)

Calculate the azimuth angle in radian from the kite position in ENU reference frame. Zero east. Positive direction clockwise seen from above. Valid range: -π .. π.

KiteUtils.ground_dist — Method

function ground_dist(vec)

Calculate the ground distance of the kite from the groundstation based on the kite position (x,y,z, z up).