Exported Types

Config files

KiteUtils.Settings — Type

Settings

Flat struct, defining the settings of the Simulator and the Viewer.

project::String

Default:

log_file::String

Default:

model::String

Default:

segments::Int64

number of tether segments Default: 0

sample_freq::Int64

Default: 0

time_lapse::Float64

Default: 0

zoom::Float64

Default: 0

fixed_font::String

Default:

v_reel_out::Float64

Default: 0

c0::Float64

Default: 0

c_s::Float64

Default: 0

c2_cor::Float64

Default: 0

k_ds::Float64

Default: 0

area::Float64

projected kite area [m^2] Default: 0

mass::Float64

kite mass incl. sensor unit [kg] Default: 0

height_k::Float64

height of the kite [m] Default: 0

alpha_cl::Vector{Float64}

Default: []

cl_list::Vector{Float64}

Default: []

alpha_cd::Vector{Float64}

Default: []

cd_list::Vector{Float64}

Default: []

rel_side_area::Float64

relative side area [%] Default: 0

alpha_d_max::Float64

max depower angle [deg] Default: 0

kcu_mass::Float64

mass of the kite control unit [kg] Default: 0

power2steer_dist::Float64

power to steering line distance [m] Default: 0

depower_drum_diameter::Float64

Default: 0

depower_offset::Float64

Default: 0

tape_thickness::Float64

Default: 0

v_depower::Float64

Default: 0

v_steering::Float64

Default: 0

depower_gain::Float64

Default: 0

steering_gain::Float64

Default: 0

v_wind::Float64

Default: 0

h_ref::Float64

Default: 0

rho_0::Float64

Default: 0

z0::Float64

Default: 0

profile_law::Int64

Default: 0

alpha::Float64

Default: 0

cd_tether::Float64

Default: 0

d_tether::Float64

Default: 0

d_line::Float64

Default: 0

height_b::Float64

height of the bridle [m] Default: 0

l_bridle::Float64

Default: 0

l_tether::Float64

Default: 0

damping::Float64

Default: 0

c_spring::Float64

Default: 0

elevation::Float64

Default: 0

sim_time::Float64

Default: 0

Also look at the default file that is parsed to fill the Settings stuct with data: settings.yaml .

Log files and internal state representation

KiteUtils.MyFloat — Type

const MyFloat = Float32

Type used for position components and scalar SysState members.

KiteUtils.SysState — Type

SysState{P}

Basic system state. One of these is saved per time step. P is the number of tether particles.

time::Float64

time since start of simulation in seconds

orient::StaticArrays.MVector{4, Float32}

orientation of the kite (quaternion, order w,x,y,z)

elevation::Float32

elevation angle in radians

azimuth::Float32

azimuth angle in radians

l_tether::Float32

tether length [m]

v_reelout::Float32

reel out velocity [m/s]

force::Float32

tether force [N]

depower::Float32

depower settings

v_app::Float32

norm of apparent wind speed [m/s]

X::StaticArrays.MVector{P, Float32} where P

vector of particle positions in x

Y::StaticArrays.MVector{P, Float32} where P

vector of particle positions in y

Z::StaticArrays.MVector{P, Float32} where P

vector of particle positions in z

KiteUtils.ExtSysState — Type

ExtSysState{P}

Extended system state. Derived values for plotting. P is the number of tether particles.

time::Float64

time since launch in seconds

orient::Rotations.QuatRotation{Float32}

orientation of the kite

x::Float32

kite position in x

y::Float32

kite position in y

z::Float32

kite position in z

KiteUtils.SysLog — Type

SysLog{P}

Flight log, containing the basic data as struct of arrays and in addition an extended view on the data that includes derived/ calculated values for plotting finally meta data like the file name of the log file is included.

name::String

name of the flight log

syslog::StructArrays.StructArray{SysState{P}} where P

struct of vectors

extlog::StructArrays.StructArray{ExtSysState{P}} where P

struct of vectors, containing derived values