Exported Types

Config files

KiteUtils.Settings — Type

Settings

Flat struct, defining the settings of the Simulator and the Viewer.

project::String

Default:

log_file::String

Default:

model::String

file name of the 3D model of the kite for the viewer Default:

physical_model::String

name of the kite model to use (KPS3 or KPS4) Default:

version::Int64

Default: 1

segments::Int64

number of tether segments Default: 0

sample_freq::Int64

Default: 0

time_lapse::Float64

Default: 0

zoom::Float64

Default: 0

kite_scale::Float64

Default: 1.0

fixed_font::String

Default:

abs_tol::Float64

Default: 0.0

rel_tol::Float64

Default: 0.0

linear_solver::String

Default: GMRES

max_order::Int64

Default: 4

max_iter::Int64

Default: 1

v_reel_out::Float64

Default: 0

c0::Float64

Default: 0

c_s::Float64

Default: 0

c2_cor::Float64

Default: 0

k_ds::Float64

Default: 0

area::Float64

projected kite area [m^2] Default: 0

mass::Float64

kite mass incl. sensor unit [kg] Default: 0

height_k::Float64

height of the kite [m] Default: 0

alpha_cl::Vector{Float64}

Default: []

cl_list::Vector{Float64}

Default: []

alpha_cd::Vector{Float64}

Default: []

cd_list::Vector{Float64}

Default: []

width::Float64

width of the kite [m] Default: 0

alpha_zero::Float64

Default: 0

alpha_ztip::Float64

Default: 0

m_k::Float64

relative nose distance; increasing m_k increases C2 of the turn-rate law Default: 0

rel_nose_mass::Float64

Default: 0

rel_top_mass::Float64

mass of the top particle relative to the sum of top and side particles Default: 0

rel_side_area::Float64

relative side area [%] Default: 0

alpha_d_max::Float64

max depower angle [deg] Default: 0

kcu_mass::Float64

mass of the kite control unit [kg] Default: 0

power2steer_dist::Float64

power to steering line distance [m] Default: 0

depower_drum_diameter::Float64

Default: 0

depower_offset::Float64

Default: 0

tape_thickness::Float64

Default: 0

v_depower::Float64

Default: 0

v_steering::Float64

Default: 0

depower_gain::Float64

Default: 0

steering_gain::Float64

Default: 0

v_wind::Float64

Default: 0

v_wind_ref::Vector{Float64}

Default: []

h_ref::Float64

Default: 0

rho_0::Float64

Default: 0

z0::Float64

Default: 0

profile_law::Int64

Default: 0

alpha::Float64

Default: 0

cd_tether::Float64

Default: 0

d_tether::Float64

Default: 0

d_line::Float64

Default: 0

h_bridle::Float64

height of the bridle [m] Default: 0

l_bridle::Float64

Default: 0

l_tether::Float64

Default: 0

damping::Float64

Default: 0

c_spring::Float64

Default: 0

rho_tether::Float64

density of Dyneema [kg/m³] Default: 0

e_tether::Float64

axial tensile modulus of the tether [Pa] Default: 0

elevation::Float64

initial elevation angle [deg] Default: 0

depower::Float64

simulation time [sim only] Default: 0

sim_time::Float64

Default: 0

temp_ref::Float64

temperature at reference height [°C] Default: 0

height_gnd::Float64

height of groundstation above see level [m] Default: 0

use_turbulence::Float64

turbulence intensity relative to Cabau, NL Default: 0

v_wind_gnds::Vector{Float64}

wind speeds at ref height for calculating the turbulent wind field [m/s] Default: []

avg_height::Float64

average height during reel out [m] Default: 0

rel_turbs::Vector{Float64}

relative turbulence at the vwindgnds Default: []

i_ref::Float64

the expected value of the turbulence intensity at 15 m/s Default: 0

v_ref::Float64

five times the average wind speed in m/s at hub height over the full year [m/s] Default: 0

height_step::Float64

grid resolution in z direction [m] Default: 0

grid_step::Float64

grid resolution in x and y direction [m] Default: 0

Also look at the default file that is parsed to fill the Settings stuct with data: settings.yaml .

Log files and internal state representation

KiteUtils.MyFloat — Type

const MyFloat = Float32

Type used for position components and scalar SysState members.

KiteUtils.SysState — Type

SysState{P}

Basic system state. One of these is saved per time step. P is the number of tether particles.

time::Float64

time since start of simulation in seconds

orient::StaticArrays.MVector{4, Float32}

orientation of the kite (quaternion, order w,x,y,z)

elevation::Float32

elevation angle in radians

azimuth::Float32

azimuth angle in radians

l_tether::Float32

tether length [m]

v_reelout::Float32

reel out velocity [m/s]

force::Float32

tether force [N]

depower::Float32

depower settings [0..1]

steering::Float32

steering settings [-1..1]

heading::Float32

heading angle in radian

course::Float32

course angle in radian

v_app::Float32

norm of apparent wind speed [m/s]

vel_kite::StaticArrays.MVector{3, Float32}

velocity vector of the kite

X::StaticArrays.MVector{P, Float32} where P

vector of particle positions in x

Y::StaticArrays.MVector{P, Float32} where P

vector of particle positions in y

Z::StaticArrays.MVector{P, Float32} where P

vector of particle positions in z

KiteUtils.SysLog — Type

SysLog{P}

Flight log, containing the basic data as struct of vectors which can be accessed as if it would be an array structs. In addition an extended view on the data that includes derived/ calculated values for plotting. Finally it contains meta data like the name of the log file.

name::String

name of the flight log

syslog::StructArrays.StructArray{SysState{P}} where P

struct of vectors that can also be accessed like a vector of structs